Moving EvidenceLevels in seperated file
